[FE] Header navigate 수정  (#262)
