+++ client/src/features/Event/Manage/pages/EventManagePage.tsx
@@ -1,10 +1,10 @@
 import { useNavigate } from 'react-router-dom';
 
+import { Button } from '@/shared/components/Button';
 import { Flex } from '@/shared/components/Flex';
 import { Header } from '@/shared/components/Header';
 import { IconButton } from '@/shared/components/IconButton';
 import { PageLayout } from '@/shared/components/PageLayout';
-import { Text } from '@/shared/components/Text';
 
 import { EventInfoSection } from '../components/EventInfoSection';
 import { GuestManageSection } from '../components/GuestManageSection';
@@ -17,18 +17,11 @@ export const EventManagePage = () => {
     <PageLayout
       header={
         <Header
-          left={
-            <Flex alignItems="center" gap="12px">
-              <IconButton
-                name="back"
-                size={14}
-                aria-label="이전 페이지로 돌아가기"
-                onClick={() => navigate(-1)}
-              />
-              <Text as="h1" type="Title" weight="semibold">
-                이벤트 관리
-              </Text>
-            </Flex>
+          left={<IconButton name="logo" size={55} onClick={() => navigate('/event')} />}
+          right={
+            <Button width="80px" size="sm" onClick={() => navigate('/event/my')}>
+              내 이벤트
+            </Button>
           }
         />
       }
