+++ client/src/features/Event/New/pages/NewEventPage.tsx
@@ -1,24 +1,26 @@
 import { css } from '@emotion/react';
 import { useNavigate } from 'react-router-dom';
 
-import { Flex } from '../../../../shared/components/Flex';
-import { Header } from '../../../../shared/components/Header';
-import { IconButton } from '../../../../shared/components/IconButton';
-import { PageLayout } from '../../../../shared/components/PageLayout';
-import { Text } from '../../../../shared/components/Text';
+import { Button } from '@/shared/components/Button';
+import { Flex } from '@/shared/components/Flex';
+import { Header } from '@/shared/components/Header';
+import { IconButton } from '@/shared/components/IconButton';
+import { PageLayout } from '@/shared/components/PageLayout';
+
 import { EventCreateForm } from '../components/EventCreateForm';
 
 export const NewEventPage = () => {
   const navigate = useNavigate();
+
   return (
     <PageLayout
       header={
         <Header
-          left={
-            <Flex alignItems="center" gap="12px">
-              <IconButton name="back" size={14} onClick={() => navigate(-1)} />
-              <Text type="Label">돌아가기</Text>
-            </Flex>
+          left={<IconButton name="logo" size={55} onClick={() => navigate('/event')} />}
+          right={
+            <Button width="80px" size="sm" onClick={() => navigate('/event/my')}>
+              내 이벤트
+            </Button>
           }
         />
       }
