+++ client/src/features/Event/Overview/pages/OverviewPage.tsx
@@ -1,32 +1,26 @@
 import { useQuery } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
 
-import { getGoogleAuthUrl, isAuthenticated } from '@/api/auth';
 import { organizationQueryOptions } from '@/api/queries/organization';
 import { Button } from '@/shared/components/Button';
 import { Flex } from '@/shared/components/Flex';
 import { Header } from '@/shared/components/Header';
-import { Icon } from '@/shared/components/Icon';
+import { IconButton } from '@/shared/components/IconButton';
 import { useGoogleAuth } from '@/shared/hooks/useGoogleAuth';
+import { theme } from '@/shared/styles/theme';
 
 import { EventList } from '../components/EventList';
 import { OrganizationInfo } from '../components/OrganizationInfo';
 
 export const OverviewPage = () => {
   const navigate = useNavigate();
+  const { logout } = useGoogleAuth();
 
   const { data: organizationData } = useQuery(
     organizationQueryOptions.organizations('woowacourse')
   );
 
   const { data: eventData } = useQuery(organizationQueryOptions.event(1));
-
-  const { logout } = useGoogleAuth();
-
-  const handleGoogleLogin = () => {
-    const authUrl = getGoogleAuthUrl();
-    window.location.href = authUrl;
-  };
 
   // S.TODO 로딩 처리
   if (!organizationData || !eventData) {
@@ -36,35 +30,11 @@ export const OverviewPage = () => {
   return (
     <>
       <Header
-        left={<Icon name="logo" width={55} />}
+        left={<IconButton name="logo" size={55} onClick={() => navigate('/event')} />}
         right={
-          <Flex gap="8px">
-            {isAuthenticated() ? (
-              <Button
-                width="80px"
-                size="sm"
-                variant="outlined"
-                onClick={logout}
-                color="#18A0FB"
-                fontColor="#18A0FB"
-              >
-                로그아웃
-              </Button>
-            ) : (
-              <Button
-                width="80px"
-                size="sm"
-                variant="filled"
-                fontColor="#FFF"
-                onClick={handleGoogleLogin}
-              >
-                로그인
-              </Button>
-            )}
-            <Button width="80px" size="sm" onClick={() => navigate('/event/my')}>
-              내 이벤트
-            </Button>
-          </Flex>
+          <Button width="80px" size="sm" onClick={() => navigate('/event/my')}>
+            내 이벤트
+          </Button>
         }
       />
       <OrganizationInfo
